crt1: exit from C world after invoking WinMain
There seem to be signals sent non-deterministically during
termination, and since Go world captures these signals
exiting the game sometimes brings about a panic stack trace.

It works about 50% of the time in Go world, and as the game
runs fine in both C and Go world, prefer C for termination so
we don't have to cross the C-Go boundary for termination.