proto: move engine events from unstable to stable
Also, remove alloc of MemLoadFile. Now the memory
of djavul.exe is usually within 60 MB. Before it was
around 300 MB at start, would grow to 600 MB after
visiting a few dungeon levels, and then crash when
loading a new tileset (as that required 400 MB to
keep track of byte locations mapping to file name,
with a very naiive implementation).

Note, prior to this commit, running the game on XP
32-bit or 32-bit wine was not possible for serious
game play, as only a few dungeon levels could be
traversed before hitting the 1.5 GB limit on wine
(3 GB on XP).